+++ common/src/main/kotlin/com/cobblemon/mod/common/api/events/pokemon/interaction/HeldItemUpdatedEvent.kt
@@ -1,0 +1,10 @@
+package com.cobblemon.mod.common.api.events.pokemon.interaction
+
+import com.cobblemon.mod.common.pokemon.Pokemon
+import net.minecraft.item.ItemStack
+
+data class HeldItemUpdatedEvent(
+    val pokemon: Pokemon,
+    val oldItem: ItemStack,
+    val newItem: ItemStack
+)